+++ src/navigation/index.js
@@ -1,5 +1,5 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import HomeScreen from './../screens/HomeScreen';
+import HomeScreen from '../screens/home';
 import ExampleScreen from './../screens/ExampleScreen';
 import NoteScreen from '../screens/note';
 import { Screens } from './constants';
