+++ src/screens/home/EmptyComponent.js
@@ -1,0 +1,10 @@
+import { Headline } from 'react-native-ios-kit';
+import React from 'react';
+
+class EmptyComponent extends React.PureComponent {
+    render() {
+        return (< Headline > Add a new note using the 'Add Note' button below</Headline >);
+    }
+}
+
+export default EmptyComponent;
