+++ src/screens/home/index.js
@@ -1,0 +1,2 @@
+import HomeScreen from "./HomeScreen";
+export default HomeScreen;
