+++ src/screens/home/styles.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        paddingTop: 44,
+        paddingHorizontal: 18,
+        backgroundColor: "white"
+    },
+    header: {
+        justifyContent: 'flex-start',
+        padding: 18,
+        paddingTop: 44,
+    },
+    body: {
+        paddingHorizontal: 16,
+    },
+    title: {
+        fontWeight: 700,
+        paddingBottom: 22,
+    },
+});
+
+export default styles;
