+++ src/screens/note/NoteTextField.js
@@ -1,7 +1,7 @@
 import { TextField } from 'react-native-ios-kit';
 import { Colors } from 'react-native-ios-kit';
 
-const NoteTextField = ({ placeholder, value, onValueChange, inputStyle, multiline }) => {
+const NoteTextField = ({ placeholder, value, onValueChange, inputStyle, multiline, ...rest }) => {
     return (<TextField
         placeholder={placeholder}
         value={value}
@@ -12,6 +12,7 @@ const NoteTextField = ({ placeholder, value, onValueChange, inputStyle, multilin
         }}
         inputStyle={inputStyle}
         multiline={multiline}
+        {...rest}
     />)
 };
